refactor: standardize API module naming to singular and add barrel exports
Renamed auth/queries.ts → query.ts and mutations.ts → mutation.ts to match the singular naming convention used by all other API modules. Renamed git/mutations.ts → mutation.ts for the same reason. Added missing barrel exports for auth/ and git/ modules, and updated the main shared/apis/index.ts to include auth, prsQueryKey, and fsQueryKey exports.
