+++ frontend/src/shared/apis/index.ts
@@ -1,6 +1,7 @@
 export { ApiClientError } from './client';
 export { projectsQuery, projectsMutation, projectsQueryKey } from './projects';
-export { prsQuery, prsMutation } from './prs';
-export { fsQuery } from './fs';
+export { prsQuery, prsMutation, prsQueryKey } from './prs';
+export { fsQuery, fsQueryKey } from './fs';
 export { chatSessionsQuery, chatSessionsQueryKey } from './chatSessions';
 export { gitMutation } from './git';
+export { authQuery, authMutation, authQueryKey } from './auth';
